requirements.txt
-e git://github.com/tothi/pyDes.git#egg=pyDes
pwntools>=4.0.1
